[kernel][phys][x86] Hand off root and identity page tables
As already done with riscv64, this is a major step toward physboot
properly creating the kernel's address space: it hands off the address
space as it is currently constructed, upper and lower, enabling the
kernel to bootstrap from the already-constructed identity map.

The similar change for arm64 is still to come.

Bug: 42164859
Change-Id: Ib32507e314438280716e4e92fe715767be8c3da9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169657
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>