[netstack3/conntrack] Remove num_connections
Instead of capping the number of connections, we cap the number of
entries in the table, which is what we actually care about. This mostly
means replacing one constant with one that's twice as big.

Since this change updated the name of the inspect value from
"num_connections" to "num_entries", also update the sampler
configuration to export that instead.

Fixed: 381946038
Change-Id: I8fc8669bba79d64ec932d6be4cf281d10ea5f453
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170952
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>