[memory] Optimize memory_monitor's `GetCapture` allocations
On devices which use the `StarnixCaptureStrategy`, there is a lot of
known memory overhead due to Starnix sharing VMOs between Linux
processes; the `Strategy` itself (which gets re-processed on every
`Digest` call, e.g. when calling `ffx profile memory`) grows quite a
lot.

This change uses the fact that an instance of `CaptureStrategy` in
practice never needs to be reused; this lets us `std::move` its
tracking state instead of `std::copying` it when consuming it, saving
the related memory.

This change saves ~15MiB of allocations on at least one device that is a
heavy user of `Starnix` processes, when profiling `memory_monitor`
during a call to `ffx profile memory`. It should both be safe and save
some CPU as well.

Change-Id: I6815b83333c0131bc4c2de26270e9faf1cb257c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168339
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: Étienne J. Membrives <etiennej@google.com>