[usb] Fork usb-endpoint-server to a SDK-clean version
We need a SDK-clean version of the usb::EndpointServer which does not
link against the DDK or banjo. This hard-forks the existing lib to a
sdk/ directory.

Note, gerrit is being stupid with the delta and showing everything as
new files rather than the git-mv it actually is. Here's what you should
be seeing, relative to src/devices/usb/lib/usb/include/usb:

1. git mv request-fidl.h internal/request-fidl.h
2. Modify internal/request-fidl.h with the templating (minor delta)
3. Add the new files request-fidl.h and sdk/request-fidl.h

Bug: b/376731071
Bug: b/376724095
Test: Tested on a lilac
Change-Id: Ia286d389bf7918daba2c0c72f170cb8b7f68d7a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167113
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>