[bazel_sdk][versioning] Handle unspecified API level after transition
Previously, fuchsia_transition failed if an API level was not specified
via any of the options. Failing in the transition prevented some
negative tests. In addition, the empty string was also handled the
FIDL rule.

As previously done for invalid API levels, allow the transition to
complete then fail when when analyzing the fuchsia_api_level rule.
The name of the relevant target is not known at this point, but the
Bazel error will indicate the target for which analysis failed.

Bug: 354047162, 382773036
Change-Id: I19ae4ef6f26ba68652ea6520c81a4690ba81cb47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170002
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>