[ulib][c] Begin soft migration of files to /strings
https://github.com/llvm/llvm-project/pull/118899 began the process of
moving files from /string to /strings. This CL likewise introduces the
/strings directory to Fuchsia's build machinery and performs the first
step in migrating the needed targets to /strings.

Since it's expected that we'll need to repeat this process as more
targets are migrated to strings/ in the future, this CL introduces a
(temporary) exec_script that can be used as an example for future CL
migrations, to make it easier and faster to detect file changes (but
note the exec_script gets cleaned up after each migration, as is done
in Ia8de544a2695125f888544071b08f37906fe381d).

Bug: 383347626
Change-Id: I81b66e1f5ec01f240ed8cf1f601764eb89ed6cff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1172248
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>