[kernel][vm] Correct reference value align bits
The extra two bits used to be used for split bit tracking directly in
the page list node reference value. This was moved to be tracked as
metadata associated with a compressed reference in the compression
storage system and so this alignment can be reduced.

Change-Id: I5b98a8b06376e518efedb9143127002162491775
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171573
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>