[debugger] Refactor Curl code to not use RefPtr
Recent changes to upstream libcurl exposed an issue with our lifetime
handling of the Curl::MultiHandle object. It is now typical to see the
curl callback functions called as a result of beginning curl's cleanup
routines.

This created an issue when tearing down the message loop, which owns
lambda functions that have the last reference to the MultiHandle object,
where more references would be created to the object from its own
destructor, and would result in reentrantly destructing, which libcurl
prevents.

The solution is a more manual management of the MultiHandle's lifetime.
There is a single object which will be destructed in the correct order
when there are no more EasyHandles associated with it. This allows the
callbacks to not require strong references to the object and for the
object to persist until all of the data has been received from the
remote end.

In the case the remote end _never_ responds to a particular request, it
is cleaned up as the first action of curl_global_cleanup.

Test: fx test --host zxdb_tests -- --gtest_filter="Curl*"
Bug: b/382444850
Change-Id: I57debface87f59f24abec1567f092358497537db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171173
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>