[ffx] Use longer knock timeout in `target wait`
Currently `ffx target wait` uses a 3-second-long knock timeout while
polling for the device to be up in `ffx target wait`. If the None
being waited on needs a security-key tap in order for SSH to succeed,
or if the target is sufficiently busy during boot, this can be too
aggressive a timeout. (This is observed in practice by noticing that
running `ffx target echo` in a loop will succeed much faster in these
scenarios than `ffx target wait` -- by around a minute.)

Address this by allowing the daemonless knock method to use its own
default timeout of 6 seconds to account for the additional setup cost.

Change-Id: Ief3494ad0e220182d1188eb088ef8aab0ab54a9e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169238
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>