[firmware] conditionally compile bootfs APIs
Passing files from the bootloader using the bootfs ZBI type was not
intended to be stable, but has become load-bearing on some legacy
devices. New devices should not use this, but the helper functions in
libzircon_boot required manually pulling in the bootfs.h definitions
(which are not part of the Firmware SDK) whether a device wanted them or
not.

This CL conditionally compiles the bootfs helpers, so that legacy boards
can define it and bring in bootfs support, but boards that don't need it
(e.g. vim3) can ignore it.

This also updates the Rust code for testing the bootfs APIs, which had
bit-rotted and needed a small update to compile.

Bug: b/378736089
Change-Id: Ief38c3788bcab08035a39e2662e41000230a2c96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173135
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>