[starnix] Send a Freeze signal with a Waiter to block the target task
When a cgroup is frozen, create a Waiter for each task in the cgroup.
Send the freeze kernel signal to each task with the Waiter to be
blocked at, waiting for an interrupt or a notification.

Handling of SIGKILL when frozen will be addressed in future CLs.

Bug: b/333766695, b/377755814, b/377755814
Change-Id: Ied9bf558a466ffb0889b440caa498da9229f0d7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173136
Commit-Queue: Guocheng Wei <guochengwei@google.com>
Reviewed-by: Adam Barth <abarth@google.com>