[sestarnix] Turn todo_check_permission() from macro to function
Encapsulate the details of a "todo_deny" check into a structure so
that the `todo_check_permission` API can be turned into a function.

The structure is still created by a macro, primarily to validate
the bug URL format, but it's a simpler one, but decoupled from the
actual check API, for greater flexibility.

Change-Id: I11f44d0d5954054ecb873b9726d6fb2176ec588c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171774
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>