[sestarnix] Introduce has_file_permission() helper
has_file_permission() accepts a FileObject and a set of file-like
permissions. It first validates that the current task has the
fd:use permission to the FileObject, before delegating to an
has_fs_node_permission() helper, which ensures that they have the
requested permissions (which may be empty) to the underlying file
node.

Bug: 364569337
Change-Id: I0935653f4d1432122100fa4f5a3fa5b80a065822
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169485
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Wez <wez@google.com>