[diagnostics] Move Starnix's stub tracking to standalone library
Starnix has a diagnostics facility for tracking "stubs" that are not
fully implemented. When these stubs are reached at runtime, the record
is recorded for inspect and potentially logged. This moves the logic to
a standalone inspect_stubs crate under the diagnostics contrib directory
so that it can be used in other contexts.

Starnix exposes this via the starnix_logging crate to the rest of the
kernel. One difference that this change introduces is that the
`starnix_logging` GN variable no longer controls logging from stubs.

Change-Id: If3cf1a57e8beef6db1dd6c92202d612564dd0011
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171241
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>