[bazel_sdk] fix analysis failure in fuchsia_api_level
This CL adds the ability to use the flag that we define for pigweed
which skips fuchsia targets to opt out of the analysis failures that
occur when we have an unset api level. This happens because in pigweed
they run builds with `bazel build //...` which ends up running the
implementation of the fuchsia_api_level rule to figure out what the
build config value should be. We cannot avoid running this implementation
so we use the flag they already have set to skip validation.

bug: 354047162
Change-Id: I8a0a3eadba5c671553355fcb1e6ac81780f60621
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173506
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>