[kernel][vm] Simplify LruThread reasoning
Tightens the coordination between the lru_event_ being signaled and the
LruThread performing processing to simplify the reasoning. The previous
logic of checking for the lru signal without the lock and performing
one processing attempt per wakeup should be correct, but attempting to
reason through a proof of this is difficult and so is worth simplyfing
the code.

Change-Id: Ie8019dc685764daf64274835fb1fca1f0f7ef2a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171874
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>