[bazel_sdk] Introduce @fuchsia_sdk//fuchsia:private_defs.bzl
Use a new file to export private details of @rules_fuchsia to the
Bazel SDK test suite. This is a first step necessary to fix the
associated bug.

Note that overriding @rules_fuchsia to point to the self-contained
SDK directory, but this fails because provider values have built-in
records of the path of the .bzl file that defined them.

This implies that `load("@first//:providers.bzl", "FooInfo")` and
`load("@second//:providers.bzl", "FooInfo")` do not return the
same value even if `@second` and `@first` point to the same content.
This results later in error at analysis time.

Without this, it would have been trivial to use
`--override_repository=@rules_fuchsia=/path/to/self_contained/fuchsia_sdk`
instead and change all load() statements in the test suite to use
`load("@rules_fuchsia//fuchsia/private:...")`.

+ Re-run all instructions in //build/bazel_sdk/tests/README.md to
  verify if they still work correctly. Add TODO links for things
  that are still broken + fix a few typos and missing arguments.

Bug: 383536158
Change-Id: I04aeffc32dba9cf75ed4c6fcd146d3766c5d208e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171835
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Turner <digit@google.com>