[fuchsia-controller] Enable logging upon construction a Context
This changes is a very basic setup for logging to a `fuchsia-controller.log` file
when initializing a Context that talks to a fuchsia device.

Note: Not every Context construction will enable logging since, in
general, initialization for a Context is skipped unless a target is
specified.

Bug: 376290078
Change-Id: Ib88202aeae5d4929abf073b520b7be85be9d42f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146612
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>