[assembly] Add dtbo support
dtbo = device tree binary overlay

Boards can supply a devicetree_overlay file, which is plumbed through
assembly and spit into the product bundle.

Partitions configs declare which partitions / slots are dedicated to the
devicetree overlays.

When `ffx target flash` is run, we flash the dtbos to the correct
partition.

When we construct an update package, we map the dtbo to `type=dtbo`, and
the paver must know how to map that to the right partition for the right
slot.

Bug: 383751674
Change-Id: I125bc932e65f5e9b7a7108473bedff5b43ce46de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173555
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>