[fint] Ensure that `fint set` calls `build/regenerator`.
This is necessary before adding any more logic to the regenerator
script. For example, [1] tries to move the generation of the Bazel
workspace to build/regenerator, but fails with an error when invoking
Bazel that says that the workspace does not exist.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1169075/3

Bug: 382019521
Change-Id: Icfeb53ab4fc87d8d720e80724c16532d1edee51a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174452
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>