[kernel][vm] Support overwriting zero intervals
Adds a function that overwrites a zero interval in one of three ways:
fully, partially from the start, or partially from the end. The new zero
interval that overwrites the old one is expected to have a different
dirty state from the old. This overwrite function gives us a more
controlled way to update dirty state bits in the intervals instead of
directly performing state changes on the sentinels.

Not used yet outside of tests.

Test: vmpl_unittest
Run-All-Tests: true
Change-Id: I6b67a2eca3e86fe1c5c434f63b5d9675ac895dbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170955
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>