[driver][power] Acquire lease to honor timeout
Make sure the system stays awake until we reach the timeout asked
for by the client, but be lazy about acquiring a wake lease. We
may obtain a wake lease when we are informed the system is
suspending instead of when the client calls `HandleInterrupt`.

Previously, if the system was resumed we didn't obtain a wake
lease. If the system started suspension later, before the
requested timeout, we would not stop it.

Change-Id: Id0b45421d42f8cf657597bbe037372c716c42970
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173498
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>