[python] Simplify self extracting binary build
The original way of doing self extraction was to embed the ZIP file to
be extracted as a data field in a Rust binary, where there is one binary
built per test. This required a full rebuild of the Rust binary for each
test change.

Two features of ELF and ZIP formats allow us to simplify the
construction of these binaries:
1. ELF files can have arbitrary data appended, which does not affect
   their execution.
2. ZIP files can have arbitrary data prepended, which does not affect
   extraction.

Instead of creating a separate binary for each test, we instead create a
base binary, and for each test we create a new copy of that binary with
the ZIP file for the test concatenated.

This speeds up local builds of the RTC test by over 65%, representing a
reduction in development cycles of over 60 seconds.

Change-Id: I58d73460fd66d9374d6bd3e37c8306ff02ec1cee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174973
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>