[netstack3/conntrack] Move make_packets to testutils
Also refactor the logic tests to use this instead of creating packets
directly.

The logic tests currently rely on the fact that self-connected flows
are immediately marked "EstablishmentLifecycle::ReplySeen", but in the
future those flows will be evictable from the table, causing the test to
fail.

Bug: 380888426
Change-Id: I5c7a90977e01a4be5b110097f2c4a45365b47f15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173792
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>