[fint] Ensure that `fint set` calls `build/regenerator`.
A reland of [1] after fixing the un-handled command-line flag that
caused the revert (+ associated regression test)

---- original commit message ----

This is necessary before adding any more logic to the regenerator
script. For example, [2] tries to move the generation of the Bazel
workspace to build/regenerator, but fails with an error when invoking
Bazel that says that the workspace does not exist.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1174452
[2] https://fuchsia-review.googlesource.com/c/fuchsia/+/1169075/3

Bug: 382019521
Change-Id: Ibde84bafa0ff80b824c6eaf027b3f11d658e3d68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175036
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Fang <fangism@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>