[dl] Introduce InitFiniArrayWithDeps test
This CL adds a test to check the order of initializers and finalizers
that are run in a module's a dependency graph in dlopen and dlclose.
This test is currently disabled on libdl.

Bug: 339302842
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I2564f06b9b544883a30b135f3bdea9027cb6f173
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1172894
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>