[fint] Add `compilation_mode` to replace `optimize`
"optimize" was originally chosen as the field name for configuring debug
vs. release builds based on the assumption that the `optimize` GN arg
would eventually be used for specifying debug vs. release but that
hasn't happened and `optimize` is used for a different purpose. So
rename the `optimize` fint parameter to `compilation_mode` to avoid
confusion with the `optimize` GN arg.

A soft transition is required to avoid breaking the infra recipes and
configs, so keep accepting `optimize` for now.

Bug: 382773442
Change-Id: Ic8a0adbdce8a86ae7afa4f3c007306063fee4835
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170792
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>