[netstack3/conntrack] Improve connection eviction
We found that connection eviction was too greedy as originally written
and would lead to connectivity problems when the table filled up. This
is always going to be a tough case, but there are a few things we can do
to improve matters.

In short, we attempt to evict a GC-eligible connection first, and then
pick the connection that's "least established". See the comment in
Table::finalize_connection for more details.

This means we'll always do a full scan for every new connection when the
table is full, but given the current table size is relatively small,
this is unlikely to be a problem in practice.

Fixed: 380888426
Change-Id: Ia6ba4ea160377a64ba81b5293b46baefab4cc6a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173793
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>