[assembly] Allowing developer overrides to override kernel args as key=value pairs
Many kernel args are in the form of 'key=value'.  This tracks
the kernel args internally in that form, so that developer
overrides can provide a new value that properly replaces the
existing one in the resultant zbi.

This also allows for the detection of conflicting values for the
same key to now be caught during assembly time.  Before, the
following two would have not been flagged as problematic:
- 'foo.bar=true'
- 'foo.bar=false'

Because these were different strings, they wouldn't be flagged
as a conflict.  Now, they'll trigger duplicate-value-for-key
errors during assembly, unless one comes from the developer
overrides.  In that case, it will always win by replacing the
value in the resultant zbi.

Bug: b/382037107
Change-Id: Ia113c43c4b4cb76c84d78b906d7ea285ad2db13b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168227
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>