[test_manager] enable debug_agent backtrace
This CL turns on test_managers integration of debug_agent, which causes
a backtrace to be sent to stderr when a test panics. This backtrace
is displayed in fx test output and is available in test output
artifacts.

Currently, when a rust test fails, we see two stack traces, because the
rust runtime also dumps the backtrace.

Test: feature is covered by current unit tests

Fixed: 286555768
Change-Id: I4c45c6746d4476d3d8c4e432578c2b408b75ae8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152601
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>