[kernel] Count mappings at phys hand-off
This change updates the arch VM backends to count the present mappings
made within the kernel address space mappings handed off from physboot
to the PMM. Currently the kernel's own boot-time MMU code accounts
for this, and so this is needed for parity (since that code is going
away).

This allows us to remove a switchback in the x86 boot MMU code to also
handle the counting of handed-off tables. This simplifies that code as
well as the related start.S logic to save the physical addresses of
some handed-off tables.

We further implement this for for riscv64, for parity and simplicity's
sake, even though wider mapping accounting is not yet in place.

Bug: 42164859
Change-Id: Icfda6f43564fca371faaa41dce059bf9593e1370
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174073
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>