[starnix] Reset the deadline in the heap after the timer is expired
`current_deadline` is checked when a new timer is added to the top
of the heap. When it's the same as the existing one, no driver action
will be taken. Android might add the same timer again after the timer
is expired. Therefore,the `current_deadline` should be cleared every
time when the timer is expired.

Bug: b/384103431
Change-Id: I43e576c156c2123fbc114cc5ba2875d03b17adef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175073
Commit-Queue: Guocheng Wei <guochengwei@google.com>
Reviewed-by: David Song <wintermelons@google.com>