[starnix] Check on the timer deadline before raising suspend signal
When the heap has only one timer and Android keeps updating the
timer with exact the same deadline, kernel will raise the signal to
prevent the suspend. We need to check on the deadline to ensure
no change is required to the driver.

Bug: b/384556971
Change-Id: I3b26ad2f07d96cb9d9f231e11d6ffb9714a438f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177573
Fuchsia-Auto-Submit: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Song <wintermelons@google.com>