[virtio-gpu-display] Use display::ModeId and display::ModeAndId.
This CL updates the C++ adapter interface to reflect the expected usage
of display modes in the Banjo/FIDL interfaces between the Display
Coordinator and engine drivers.

The implementations hard-code a ModeId of 1, or expect a ModeId of 1.
This matches the expected implementations of engine drivers that support
a single preferred mode.

Bug: 384573759
Test: fx test //src/graphics/display/drivers/virtio-gpu-display
Test: Booted workbench_eng.x64 on qemu
Test: ffx target ssh display-tool squares
Change-Id: Ib7ef5d4b1dde0efad35c021f8d678eb1e71cc1c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1176095
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>