[audio][drivers][aml-g12] Calculate VMO size from requested format
`GetVmo` calculates the required VMO size from the requested
minimum number of frames and the ring buffer format.

Previously, it used the format set in `engine_config.ring_buffer`
that is set in `ResetEngine` from the DAI format.

This happens to work if the client first sets a DAI format that
matches the requested ring buffer format passed to `CreateRingBuffer`,
but calculates the wrong size in two cases:

1. Client calls GetVmo before setting DAI format
2. Client uses a ring buffer that is supported, but does not match
   the previously set (or default) DAI format

Case #1 is useful to support because it allows the client to power
the SoC on using `SetActiveChannels` before setting the DAI format
so the call takes effect.

ADR calls `GetVmo` once the ring buffer is created. Without this
change, ADR rejects the VMO because the VMO size is too small.

Bug: b/384110895
Change-Id: Iac4117eae21dbee57721c62e81edd2af40447b03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175802
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>