[feedback] Add fuchsia-page_fault-timed_out signature
There's not much the program can do if the pager is timing out on
delivering the requested page to the program. This CL will make it so
that the signature in this situation is fuchsia-page_fault-timed_out
instead of blaming the program that made the request.

Bug: 363257341
Tested: `fx test //src/developer/forensics`
Change-Id: I4d8fc1ad38b58b40d6e5389a784fd3b7100bc89e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178674
Commit-Queue: Thomas Gales <tgales@google.com>
Reviewed-by: Alex Pankhurst <pankhurst@google.com>