[dl] libdl can run initializers
Introduce `RuntimeDynamicLinker::RunInitializers` that takes the module
returned by a successful `Open` result and runs initializers on that
module and its dependencies. Initializers are run in reverse order of
the dependency graph (i.e. the reverse load order) so that init funcs
for dependencies are run before the dependent module.

This does not include support for synchronization or handle re-entrency,
which will be addressed in a future change.

Fixed: 382340693
Change-Id: I74616517da5aafb29ae82acdae458e33c5c96900
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174713
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>