[kernel][phys] Drop to EL1 early in physload
We drop to EL1 - disabling any EL2 monitor - early in phyload, before
setting up the address space. This simplifies the hand-off to EL1, as
then we'll set up the address space in EL1 as opposed to EL2 and don't
have to futz with arm-specific reinstallation. This is prudent also in
that the kernel proper no longer executes in EL2 and we have no need to
do so in physload/physboot.

Bug: 42164859
Change-Id: I9f27dc3f9e781c28f583e16816e57f12561f3841
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1176696
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>