[ffx][emu] Assign host ports when using qemu
This adds assigning the host port to a port that is forwarded out of the
guest. Older versions of QEMU automatically assigned an unused port to
complete the mappings, however QEMU 9 does not do this and so
ffx does this as part staging the instance.

Bug: 384957072
Change-Id: I2f9ad920459e1c4dfa6caa781a9f47aceb1c4225
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178082
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>