[assembly] Stop adding build-info/minimum-utc-stamp
We have two backstop files:

  /config/build-info/minimum-utc-stamp is supplied by the product,
  but is never used.

  /boot/config/build_info/minimum_utc_stamp is supplied by the platform,
  and is used as the backstop.

We are removing support for the former, because it is unused. The first
step is to make sure the file supplied by the product is not added to
the final images. Then we can remove the usages in product, and remove
the config field altogether.

Bug: 356621614
Change-Id: If71e9494eeaf2c3ef26cbf480406fdf6e0a6368e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179433
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>