[virtio-gpu-display] Fix incorrect cleanup during driver shutdown.
Due to a typo, DisplayEngineEventsBanjo::SetListener() does not
correctly reset the listener when given a nullptr. The bug results in
event listeners still being called. This CL fixes the bug, to address
any false rejections in CI/CQ that may be caused by this problem.

In the future, the Banjo adapters will be lifted into a library with
full test coverage. This bug was discovered by the work for that
project.

Bug: 42079190
Test: none (test coverage will be provided in future CLs)
Change-Id: I81d4660b041e12537711efa3ae37c30484b938ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178539
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Yilong Li <liyl@google.com>