[ffx] Refactor ffx_plugin_impl to use fhoenvironment
This removes the direct use of the injector from the main
frontend of ffx. This was  legacy code from supporting the old
plugin architecture.

This change adds using the FhoEnvironment which is part of the new,
current subtool architecture.

This also removes the deprecated injector field in FhoEnvironment.

Change-Id: I988b15d947042a6738606107cb952ea2ade7ed81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180440
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>