[ffx][doctor] Removing dep on ffx_command and global Env
This removes the dependency on ffx_command, which was used to
get the target spec. It was replaced with calling
ffx_target_get_target_spec.

Since this call needs an EnvironmentContext, this was added to
the FfxTool struct, and the global context usages replaced.

Change-Id: Ifb2143f964ecedb7cdcd6ec12da515a87bf6fa45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181013
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>