[ffx] Use ffx_command_error vs. re-exported
This changes usages of fho::Error, Result and related macros
and usages of ffx_command::Error, Result to use ffx_command_error
directly.

This is in preparation of refactoring the internals of ffx and
making clear visibility GN rules for what defines "internal to ffx"
and APIs that should be used by subtools.

Change-Id: I93dfc713571c4f091f39679ed9448441fcafa837
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180342
Reviewed-by: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>