[display] Create standalone protocol adapters library.
This CL extracts the adapters in the virtio-gpu-display driver. The code
is similar to the driver, with one exception -- the
DisplayEngineBanjoAdapter class became a template, so it can accommodate
different DisplayEngineInterface implementations.

This CL has fairly empty tests for the adapter implementations. Unit
tests will be fleshed out in separate CLs, so we can iterate on the
approach to mocking.

Bug: 42079190
Test: fx test //src/graphics/display/lib/api-protocols
Change-Id: I7f4dfecc49ce59e185e3e79655443c2021c67d34
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178735
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>