[build] Add dir parameter to golden_files()
When set, this checks that the directory contains nothing but
exactly the set of golden files for the target.  This will
facilitate more indirect and/or automated means of maintaining
the lists of golden files and while avoiding any old unused files
collecting in the source tree.

Bug: 376333113
Change-Id: I1009e0951eec06c7ae2e359655001571ed769e3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180612
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>