[bazel_sdk] move fuchsia_targets_enabled to rules_fuchsia
This moves the definition of fuchsia_targets_enabled to its own location
that only exists in rules_fuchsia. This avoids the flag being available
on both rules_fuchsia and fuchsia_sdk which is error prone. The change
leaves an alias in place to facilitate a soft transition.

bug: 386845692
Change-Id: I462331444a2c9eca39b6be75c644325067c62d65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181024
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>