[starnix] CloneHelper only inspects errno on MAP_FAILED.
The assertion that errno == 0 has been causing occasional flakes in
the waitpid_flags test suite. However it's not guaranteed that
errno will be equal to 0 when mmap succeeds, so we should not assert
on its value. It seems likely that the flake is caused by a previous
test leaving a non-zero errno value before this code executes.

If the mmap is actually failing, then we will still detect that by
checking for MAP_FAILED and logging errno's value if that's the case.

For now, mark the flake as fixed and let it get re-filed if the mmap
is actually failing to cause the flake.

Fixed: 387167196
Change-Id: I2918de677450bd839b0ec2d6bd955d2b5787a8f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181772
Reviewed-by: Fabio D'Urso <fdurso@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>